Problem: Needed some of the fixtures from `tests/upsert_validator/conftest.py` for tests in another directory
Solution: Moved the fixtures to `tests/conftest.py`
